apis/image_gen_api/requirements.txt
torch>=2.1.0
diffusers>=0.32.0
transformers>=4.45.0
accelerate>=0.25.0
safetensors>=0.4.0
sentencepiece>=0.2.0
peft>=0.13.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
pillow>=10.0.0
